Unify get table location in tests
