+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeBasic.java
@@ -56,13 +56,13 @@ public class TestDeltaLakeBasic
     public void registerTables()
     {
         for (String table : Iterables.concat(PERSON_TABLES, OTHER_TABLES)) {
-            String dataPath = getTableLocation(table).toExternalForm();
+            String dataPath = getResourceTableLocation(table).toExternalForm();
             getQueryRunner().execute(
                     format("CALL system.register_table('%s', '%s', '%s')", getSession().getSchema().orElseThrow(), table, dataPath));
         }
     }
 
-    private URL getTableLocation(String table)
+    private URL getResourceTableLocation(String table)
     {
         return getClass().getClassLoader().getResource("databricks/" + table);
     }
@@ -122,7 +122,7 @@ public class TestDeltaLakeBasic
         // create a bad_person table which is based on person table in temporary location
         String tableName = "bad_person";
         Path tableLocation = Files.createTempFile("bad_person", null);
-        copyDirectoryContents(Path.of(getTableLocation("person").toURI()), tableLocation);
+        copyDirectoryContents(Path.of(getResourceTableLocation("person").toURI()), tableLocation);
         getQueryRunner().execute(
                 format("CALL system.register_table('%s', '%s', '%s')", getSession().getSchema().orElseThrow(), tableName, tableLocation));
