+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/BaseTestHiveOnDataLake.java
@@ -1842,9 +1842,4 @@ public abstract class BaseTestHiveOnDataLake
                                 PartitionStatistics.empty()))
                         .collect(toImmutableList()));
     }
-
-    private String getTableLocation(String tableName)
-    {
-        return (String) computeScalar("SELECT DISTINCT regexp_replace(\"$path\", '/[^/]*$', '') FROM " + tableName);
-    }
 }
